int-> unsigned short

+++ main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <Windows.h>
 
-void setcolor(int color){
+void setcolor(unsigned short color){
     HANDLE hcon=GetStdHandle(STD_OUTPUT_HANDLE);
     SetConsoleTextAttribute(hcon, color);
 }
